Move the InMemory artifact to exporters.
